dentist_booking/requirements.txt
asgiref==3.5.2
Django==3.2
django-admin-rangefilter==0.8.4
django-background-tasks==1.2.5
django-compat==1.0.15
django-phonenumber-field==6.3.0
phonenumbers==8.12.50
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
